% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_data_functions.R
\name{byRow}
\alias{byRow}
\title{function to concatenate a matrix into a vector by row}
\usage{
byRow(mat)
}
\arguments{
\item{mat}{}
}
\description{
function to concatenate a matrix into a vector by row
}
