% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{cvd_patient}
\alias{cvd_patient}
\alias{cvd_geno}
\title{CVD risk synthetic dataset}
\format{An object of class \code{data.frame} with 425195 rows and 13 columns.}
\source{

}
\usage{
data(cvd_patient)
}
\description{
A synthetic dataset for predicting cardiovascular risk in patient cohorts.
}
\details{

}
\examples{
data(cvd_patient)

#look at summary of data
summary(cvd_patient)
}
\keyword{datasets}
