% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_data_functions.R
\name{femaleRiskScore}
\alias{femaleRiskScore}
\title{Given generated data, calculate framingham risk score for females}
\usage{
femaleRiskScore(df)
}
\arguments{
\item{df}{}
}
\description{
Given generated data, calculate framingham risk score for females
}
