% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_data_functions.R
\name{maleRiskScore}
\alias{maleRiskScore}
\title{Given generated data, calculate framingham risk score for Males}
\usage{
maleRiskScore(df)
}
\arguments{
\item{df}{}
}
\value{
original data.frame with extra column outProb
}
\description{
Given generated data, calculate framingham risk score for Males
}
