% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_data_functions.R
\name{selectInRange}
\alias{selectInRange}
\title{Resample within a category}
\usage{
selectInRange(rangeString)
}
\arguments{
\item{rangeString}{}
}
\description{
convenience function for resampling within a range of values
i.e., if category is 10-20, then will select a value between
10 and 20.
}
