% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_data_functions.R
\name{transformDataSet}
\alias{transformDataSet}
\title{Given categorical data, generate risk score}
\usage{
transformDataSet(testData)
}
\arguments{
\item{testData}{}
}
\description{
Given categorical data, generate risk score
}
