% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{cvd_patient}
\alias{cvd_patient}
\alias{cvd_genodata}
\title{CVD risk synthetic dataset}
\format{a data frame with variables related to cvd risk. \code{cvd_genodata} is a smaller subset
  of the data with a few genetic cohorts.}
\usage{
data(cvd_patient)
}
\description{
A synthetic dataset for predicting cardiovascular risk in patient cohorts.
}
\details{
\itemize{
    \item patientID Patient Identifier. Unique patient identifier for Health Hospital University.
    HHUID + 8 digit code
    \item age Patient Age Category. Age category of patient. string of age ranges.
    \item htn Does patient have hypertension? Threshold systolic blood pressure is 150. Y/N
    \item treat Is patient receiving hypertension treatment? Y/N
    \item smoking Y/N based on threshold of pack years: 10
    \item race Race based on self-defined question in survey. AmInd (american indian),
    Asian/PI (asian/pacific islander), Black/AfAm (Black/African American), White
    \item gender Gender of patient. Male, Female, NA means that patient did not want gender recorded.
    \item t2d Whether patient has Type 2 diabetes. Y/N.
    \item numAge numerical age in years.
    \item bmi Body Mass Index of Patient. kg/m^2
    \item sbp Systolic Blood Pressure in mm/Hg
    \item cvd Cardiovascular disesase based on extraction from patient billing codes. Y/N
  }

There is a subset of patients that include genetic covariate information called \code{cvd_genodata}.

\itemize{
  \item rs10757278 SNP data
  \item rs1333049 SNP data
  \item rs4665058 SNP data
  \item rs8055236 SNP data
 }
}
\examples{
#load full dataset
data(cvd_patient)

#look at summary of data
summary(cvd_patient)

#load genotype dataset
data(cvd_genodata)
summary(cvd_genodata)
}
\keyword{datasets}
